requirements.txt
Flask
Flask-Cors
requests
python-dotenv
websocket-client 